Moved cleanup logic into seperate file
